Fix test errors and introduce Session to AuthenticationContext
