+++ vertx-web-auth-common/src/main/java/io/vertx/ext/auth/common/AuthenticationContext.java
@@ -35,4 +35,6 @@ public interface AuthenticationContext {
     // NOOP
   }
 
+  Session session();
+
 }
