+++ vertx-web/src/main/java/io/vertx/ext/web/handler/impl/WebAuthenticationHandlerImpl.java
@@ -3,10 +3,8 @@ package io.vertx.ext.web.handler.impl;
 import io.vertx.core.http.HttpHeaders;
 import io.vertx.ext.auth.User;
 import io.vertx.ext.auth.authentication.AuthenticationProvider;
-import io.vertx.ext.auth.common.UserContextInternal;
 import io.vertx.ext.auth.common.handler.impl.AuthenticationHandlerImpl;
 import io.vertx.ext.web.RoutingContext;
-import io.vertx.ext.web.Session;
 import io.vertx.ext.web.common.HttpException;
 
 public abstract class WebAuthenticationHandlerImpl<T extends AuthenticationProvider> extends AuthenticationHandlerImpl<RoutingContext, T> {
@@ -20,14 +18,6 @@ public abstract class WebAuthenticationHandlerImpl<T extends AuthenticationProvi
   }
 
   public void postAuthentication(RoutingContext ctx, User authenticatedUser) {
-    ((UserContextInternal) ctx.user())
-      .setUser(authenticatedUser);
-    Session session = ctx.session();
-    if (session != null) {
-      // the user has upgraded from unauthenticated to authenticated
-      // session should be upgraded as recommended by owasp
-      session.regenerateId();
-    }
     ctx.next();
   }
